demos/02-study-case-anthropic-tools-resources-prompts-chat-app/requirements.txt
# MCP (Model Context Protocol)
mcp>=1.0.0

# OpenAI API
openai>=1.0.0

# CLI Interface
prompt-toolkit>=3.0.0

# Environment variables
python-dotenv>=1.0.0

# JSON handling (usually included in Python standard library)
# pydantic (dependency of mcp)